xiaoyaopeipei-user-merchant-backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.5
alembic==1.13.1

pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.10

dashscope==1.14.1
openai==1.12.0

oss2==2.18.4
pillow==10.2.0
# 阿里云 STS SDK（用于生成临时 OSS 访问凭证）
aliyun-python-sdk-sts==3.1.3
aliyun-python-sdk-core==2.16.0
jmespath>=0.9.3,<1.0.0

cachetools==5.3.3

qrcode[pil]==7.4.2

python-dotenv==1.0.1
pydantic-core==2.16.3
annotated-types==0.6.0

httpx==0.27.0
requests==2.31.0

python-dateutil==2.9.0.post0

loguru==0.7.2

typing-extensions==4.10.0

pyyaml==6.0.1

pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

flake8==7.0.0
black==24.2.0
isort==5.13.2
mypy==1.8.0

types-requests==2.31.0.20240218
types-python-dateutil==2.8.19.20240106

ipython==8.22.1
ipdb==0.13.13
